requirements.txt
streamlit
langchain
langchain-community
langchain-groq
faiss-cpu
python-dotenv
requests
beautifulsoup4
sentence-transformers
tiktoken